+++ NEWS.md
@@ -4,6 +4,12 @@
   than `origin`, following in `stat_bin()`'s footsteps (@teunbrand).
 * `stat_summary_2d()` and `stat_bin_2d()` now deal with zero-range data
   more elegantly (@teunbrand, #6207).
+* `geom_ribbon()` now appropriately warns about, and removes, missing values 
+  (@teunbrand, #6243).
+* `guide_*()` can now accept two inside legend theme elements:
+  `legend.position.inside` and `legend.justification.inside`, allowing inside
+  legends to be placed at different positions. Only inside legends with the same
+  position and justification will be merged (@Yunuuuu, #6210).
 * New stat: `stat_manual()` for arbitrary computations (@teunbrand, #3501)
 * Reversal of a dimension, typically 'x' or 'y', is now controlled by the 
   `reverse` argument in `coord_cartesian()`, `coord_fixed()`, `coord_radial()`